refactor(car): separate frontend and backend tool surfaces
