+++ examples/car/.env.example
@@ -12,6 +12,8 @@ QWEN_AUDIO_REALTIME_VOICE=longanqian
 COCKPIT_DOMAIN_HOST=127.0.0.1
 COCKPIT_DOMAIN_PORT=3010
 COCKPIT_DOMAIN_ORIGIN=http://127.0.0.1:3010
+# Optional explicit foreground MCP endpoint; derived from COCKPIT_DOMAIN_ORIGIN by default.
+# COCKPIT_FRONTEND_MCP_URL=http://127.0.0.1:3010/mcp/frontend
 COCKPIT_AGENT_HOST=127.0.0.1
 COCKPIT_AGENT_PORT=3020
 COCKPIT_AGENT_CARD_URL=http://127.0.0.1:3020/.well-known/agent-card.json
