+++ examples/car/README_ZH.md
@@ -45,8 +45,9 @@ npm run example:car
 - UI 仅通过 GCP 与 Gateway 对话，不感知 Realtime Provider 或后台 Agent。
 - 主座舱区域保持纯语音交互；文字转写只进入调试面板，并且 ASR 仅展示最终结果。
 - UI 通过场景自己的 HTTP/SSE 通道展示车辆、路线、音乐、天气和订单状态，以及细粒度场景进度；Gateway 不解析这些对象。
-- 前台 Agent 负责实时聊天，座舱任务通过固定的 `spawn_thinking` 桥梁提交给后台；场景只在 [`spawn-thinking-tool.mjs`](spawn-thinking-tool.mjs) 定义后台能力描述，不改工具名称和参数协议。
-- 示例后台通过 A2A 接入 Gateway，并通过 MCP 调用领域能力；它只实现少量意图路由，不模拟完整行业 Agent。
+- 前台 Agent 负责实时聊天，可通过标准 MCP 直接调用只读天气工具；其他座舱任务通过固定的 `spawn_thinking` 桥梁提交给后台。
+- 示例后台通过 A2A 接入 Gateway，并通过独立的后台 MCP 工具面调用车控、导航、音乐和闪购；它只实现少量意图路由。
+- 场景工具按领域收敛在 [`tools/`](tools/README.md)，开发者通过显式注册表增加工具包或调整前后台归属，不需要修改 Gateway 协议。
 - 客户可以替换整个 UI、后台 Agent 或领域服务，而不修改框架核心。
 
 ## 开发与测试
