+++ examples/car/frontend-profile.json
@@ -2,5 +2,8 @@
   "version": 1,
   "name": "cockpit-example",
   "description": "A concise, action-oriented foreground voice assistant for the cockpit example.",
-  "assistant": "ASSISTANT.md"
+  "assistant": "ASSISTANT.md",
+  "toolSources": {
+    "mcp": "tools/frontend-mcp.json"
+  }
 }
