+++ examples/car/spawn-thinking-tool.mjs
@@ -1,5 +1,5 @@
 export const COCKPIT_SPAWN_THINKING_DESCRIPTION = [
   '将需要座舱 Agent 执行的任务异步提交到后台。',
-  '用于车辆车窗、天窗、灯光和空调控制，导航与当前路线查询，音乐播放，天气查询，以及示例闪购流程。',
+  '用于车辆车窗、天窗、灯光和空调控制，导航与当前路线查询，音乐播放，以及示例闪购流程。',
   '不要用它处理普通闲聊，也不要声称它具备桌面文件、屏幕、代码或其他未列出的能力。',
 ].join('')
