+++ examples/car/tools/flashbuy/manifest.json
@@ -3,37 +3,25 @@
   "label": "淘宝闪购",
   "description": "座舱示例的附近商品搜索、加购、订单预览、确认下单和取消能力。",
   "enabled": true,
-  "routeRules": [
-    "用户表达任何外卖、奶茶、咖啡、点餐、淘宝闪购、下单相关意图 → 必须调用 flashbuy",
-    "用户只说“看看/搜一下/有哪些”时 → 调用 flashbuy(action=search)",
-    "用户说“帮我点/来一杯/点杯/想喝/想吃”时 → 优先调用 flashbuy(action=add_to_cart)，由 executor 搜索、选择候选、加入购物车并给出订单预览，然后询问是否确认下单",
-    "只有已经给用户播报或展示过订单预览后，用户再次明确说“确认”“下单”“买”“就这个”“可以”等确认意图时，才调用 flashbuy(action=confirm_order, confirmed=true)",
-    "如果当前还没有订单预览，用户说“确认”“下单吧”时，禁止直接 confirm_order；应先调用 flashbuy(action=add_to_cart 或 preview_order) 生成预览，再询问确认",
-    "下单前必须让用户听到或看到商品、价格、送达位置和预计送达时间"
-  ],
-  "examples": [
-    "用户说“帮我点杯奶茶” → 调用 flashbuy(action=add_to_cart, query=奶茶)",
-    "用户说“看看附近有什么外卖” → 调用 flashbuy(action=search, category=food)",
-    "用户说“确认下单”且已有订单预览 → 调用 flashbuy(action=confirm_order, confirmed=true)"
-  ],
   "functions": [
     {
       "name": "flashbuy",
       "label": "淘宝闪购",
       "description": "搜索外卖/奶茶/咖啡等附近可配送商品、加入购物车、修改购物车、预览订单、确认下单或取消。用户说帮我点/来一杯/想喝/想吃时优先 add_to_cart 生成订单预览；用户只是看看/搜一下/有哪些时用 search。下单前必须先预览订单并获得用户明确确认；没有订单预览时禁止直接 confirm_order。",
-      "executor": "flashbuy.run",
       "enabled": true,
-      "exposure": {
-        "text": true,
-        "voice": true,
-        "debug": true
-      },
       "parameters": {
         "type": "object",
         "properties": {
           "action": {
             "type": "string",
-            "enum": ["search", "add_to_cart", "update_cart", "preview_order", "confirm_order", "cancel_order"],
+            "enum": [
+              "search",
+              "add_to_cart",
+              "update_cart",
+              "preview_order",
+              "confirm_order",
+              "cancel_order"
+            ],
             "description": "search=搜索商品, add_to_cart=加入购物车, update_cart=修改购物车, preview_order=预览订单, confirm_order=确认下单, cancel_order=取消当前闪购流程"
           },
           "query": {
@@ -42,7 +30,10 @@
           },
           "category": {
             "type": "string",
-            "enum": ["food", "tea"],
+            "enum": [
+              "food",
+              "tea"
+            ],
             "description": "food=外卖热餐, tea=奶茶饮品"
           },
           "itemId": {
@@ -66,7 +57,9 @@
             "description": "只有用户明确说确认、下单、买、支付等确认意图时才为 true"
           }
         },
-        "required": ["action"]
+        "required": [
+          "action"
+        ]
       }
     }
   ]
