+++ examples/car/tools/music/catalog.mjs
@@ -1,0 +1,8 @@
+export const SONGS = Object.freeze([
+  Object.freeze({ id: 'sunny-day', title: '晴天', artist: '周杰伦', album: '叶惠美' }),
+  Object.freeze({ id: 'common-jasmine-orange', title: '七里香', artist: '周杰伦', album: '七里香' }),
+  Object.freeze({ id: 'rice-field', title: '稻香', artist: '周杰伦', album: '魔杰座' }),
+  Object.freeze({ id: 'nocturne', title: '夜曲', artist: '周杰伦', album: '十一月的萧邦' }),
+  Object.freeze({ id: 'simple-love', title: '简单爱', artist: '周杰伦', album: '范特西' }),
+  Object.freeze({ id: 'blue-and-white-porcelain', title: '青花瓷', artist: '周杰伦', album: '我很忙' }),
+])
