+++ examples/car/tools/navigation/catalog.mjs
@@ -1,0 +1,5 @@
+export const DEFAULT_ORIGIN = Object.freeze({
+  name: '阿里巴巴云谷园区',
+  location: '120.037239,30.318522',
+  city: '杭州',
+})
