+++ examples/car/tools/navigation/manifest.json
@@ -3,29 +3,12 @@
   "label": "导航",
   "description": "座舱示例的地点搜索、路线规划、开始导航与停止导航能力。",
   "enabled": true,
-  "routeRules": [
-    "用户表达开始导航、带我去、导航到某地 → 必须调用 navigation_start",
-    "用户表达查路线、看看怎么走、规划路线但未要求开始导航 → 必须调用 navigation_route_query",
-    "用户表达停止导航、结束导航、取消路线 → 必须调用 navigation_stop",
-    "导航是最高优先级任务，应立即响应，不要反复确认"
-  ],
-  "examples": [
-    "用户说“导航到西湖” → 调用 navigation_start(destination=西湖)",
-    "用户说“看看去西湖怎么走” → 调用 navigation_route_query(destination=西湖)",
-    "用户说“停止导航” → 调用 navigation_stop()"
-  ],
   "functions": [
     {
       "name": "navigation_start",
       "label": "开始导航",
       "description": "搜索目的地、规划驾车路线并开始导航。",
-      "executor": "navigation.start",
       "enabled": true,
-      "exposure": {
-        "text": true,
-        "voice": true,
-        "debug": true
-      },
       "parameters": {
         "type": "object",
         "properties": {
@@ -39,24 +22,28 @@
           },
           "strategy": {
             "type": "number",
-            "enum": [0, 13, 5, 4, 11, 14, 2],
+            "enum": [
+              0,
+              13,
+              5,
+              4,
+              11,
+              14,
+              2
+            ],
             "description": "路线偏好: 0=智能推荐, 13=高速优先, 5=不走高速, 4=躲避拥堵, 11=少收费, 14=大路优先, 2=时间优先。当用户提及路线偏好时设置此参数"
           }
         },
-        "required": ["destination"]
+        "required": [
+          "destination"
+        ]
       }
     },
     {
       "name": "navigation_route_query",
       "label": "路线查询",
       "description": "搜索目的地并规划驾车路线，但不开始导航。",
-      "executor": "navigation.route_query",
       "enabled": true,
-      "exposure": {
-        "text": true,
-        "voice": true,
-        "debug": true
-      },
       "parameters": {
         "type": "object",
         "properties": {
@@ -70,7 +57,15 @@
           },
           "strategy": {
             "type": "number",
-            "enum": [0, 13, 5, 4, 11, 14, 2],
+            "enum": [
+              0,
+              13,
+              5,
+              4,
+              11,
+              14,
+              2
+            ],
             "description": "路线偏好: 0=智能推荐, 13=高速优先, 5=不走高速, 4=躲避拥堵, 11=少收费, 14=大路优先, 2=时间优先。当用户提及路线偏好时设置此参数"
           }
         }
@@ -80,13 +75,7 @@
       "name": "navigation_stop",
       "label": "停止导航",
       "description": "停止当前导航并清空地图路线。",
-      "executor": "navigation.stop",
       "enabled": true,
-      "exposure": {
-        "text": true,
-        "voice": true,
-        "debug": true
-      },
       "parameters": {
         "type": "object",
         "properties": {}
